+++ functional-test/src/test/java/lib/SecretGeneratorInvalidReferenceHandlingTest.java
@@ -7,9 +7,9 @@ import ch.qos.logback.core.read.ListAppender;
 import dev.zucca_ops.kustomtrace.KustomTrace;
 import dev.zucca_ops.kustomtrace.model.KustomGraph;
 import dev.zucca_ops.kustomtrace.model.Kustomization;
+import dev.zucca_ops.kustomtrace.parser.ReferenceExtractors;
 import org.junit.jupiter.api.Test;
 import org.slf4j.LoggerFactory;
-import dev.zucca_ops.kustomtrace.parser.ReferenceExtractors;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -17,17 +17,15 @@ import java.nio.file.Paths;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class ConfigMapInvalidReferenceHandlingTest {
+public class SecretGeneratorInvalidReferenceHandlingTest {
 
     ListAppender<ILoggingEvent> listAppender = new ListAppender<>();
     Kustomization app;
 
     private void assertInvalidReference(String file) {
-        // Assert that no dependencies of app contain the invalid parts in their paths
         assertThat(app.getDependencies())
                 .noneMatch(path -> path.toString().contains(file));
 
-        // Assert that invalid references were logged
         assertThat(listAppender.list)
                 .anySatisfy(event -> {
                     assertThat(event.getLevel()).isEqualTo(Level.WARN);
@@ -37,11 +35,10 @@ public class ConfigMapInvalidReferenceHandlingTest {
                 });
     }
 
-
     @Test
-    void testInvalidConfigmapReferences() throws IOException {
+    void testInvalidSecretGeneratorReferences() throws IOException {
         Logger logger = (Logger) LoggerFactory.getLogger(ReferenceExtractors.class);
-        Path appsDir = Paths.get( "src", "test", "resources", "all-reference-types-apps", "app-configmap");
+        Path appsDir = Paths.get("src", "test", "resources", "all-reference-types-apps", "app-secret");
         listAppender.start();
         logger.addAppender(listAppender);
         KustomTrace kustomTrace = KustomTrace.fromDirectory(appsDir);
@@ -50,7 +47,7 @@ public class ConfigMapInvalidReferenceHandlingTest {
         Path appPath = appsDir.resolve("kustomization.yaml");
         app = graph.getKustomization(appPath);
 
-        assertInvalidReference("non-existing-configmap.yaml");
+        assertInvalidReference("non-existing-secret.txt");
         assertInvalidReference("invalid-directory");
 
         assertThat(app.getDependencies())